sagemaker-deploy/requirements.txt
onnxruntime-gpu==1.16.0
numpy==1.24.3
